Fix automatic pythonpath modification if invoked from source tree
Assuming we have checked out rubber under ~/code/rubber, we previously
added ~/code/rubber/rubber to the pythonpath, when it should be
~/code/rubber instead.

We further added the Path object, when we should instead add Strings
to sys.path (it appears everything else is ignored?).

This furthermore ensures that we add the path only once. Even though
the scripts in bin/ are usually run once, they may be repeatedly
invoked by the same Python interpreter if run within PDB.