[latex] Fix that no error details are shown if generating command fails
The commit a164332 ("Move execute from Environment class to util.py.")
changed compile() in latex.py so that it would return False as soon as
the command, e.g. pdflatex, returned a non-zero exit
code. Unfortunately, this means that parse_log() was no longer
invoked, which caused Rubber to stay silent about the cause of the
non-zero exit status.

To fix this, we now call always parse_log() after execute(), and check
the exit code only after that.

Fixes: a164332 ("Move execute from Environment class to util.py.")