[cmdline] Remove unnecessary and erroneously check for --jobname
This check, which was introduced with 9009066 ("Parse arguments with
argparse instead of getopt."), is erroneously because it is also
performed if rubber-pipe is invoked. But with rubber-pipe we can have
--jobname *without* any source files specified.

This leads to the following traceback with e.g. "rubber-pipe --jobname
test":

Traceback (most recent call last):
  File "/usr/lib/python-exec/python3.7/rubber-pipe", line 18, in <module>
    rubber.cmdline.main (command_name = rubber.cmdline.RUBBER_PIPE)
  File "/usr/lib/python3.7/site-packages/rubber/cmdline.py", line 282, in main
    options = parse_opts (command_name)
  File "/usr/lib/python3.7/site-packages/rubber/cmdline.py", line 225, in parse_opts
    if args.jobname is not None and 1 < len (args.source):
AttributeError: 'Namespace' object has no attribute 'source'

The check is furthermore unnecessary because we already set nargs='+'
in the argument definition of 'source'. Hence argparse will already
bail out if no source was provided.

Fixes: 9009066 ("Parse arguments with argparse instead of getopt.")