[build] Extract version from first version entry in NEWS file
Instead of expecting that the first line of the NEWS file always
starts with "Version", we now wind down to the first line that starts
with "Version".

This allows the NEWS file to contain an entry with a list of
high-level end-user-friendly changes for the upcoming version (of
which the exact version number is often unknown).